% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildENPdatasets.R
\name{BuildENPdatasets}
\alias{BuildENPdatasets}
\title{BuildENPdatasets function}
\usage{
BuildENPdatasets(dat=NULL, mn.el=FALSE, ineq.ind="Gini",
                        filterIndependents=FALSE, filterSingleConstituencies=TRUE)
}
\arguments{
\item{dat}{dataset "CLEA Lower/Upper Chamber Elections Archive". If omitted, the script downloads data from the CLEA website.}

\item{mn.el}{adds month to the index variable}

\item{ineq.ind}{changes inequality index (by default computes indices based on "Gini"; for more detail about other indices see ineq() package)}

\item{filterIndependents}{filters out independent candidates (by default filterIndependents=FALSE, i.e. includes independents)}

\item{filterSingleConstituencies}{filters out single-member constituencies (by default filterSingleConstituencies=FALSE,
i.e. includes single constituencies)}
}
\value{
A list with the following parameters:
\itemize{
  \item national.level.data  -  national nationalization dataset
  \item party.level.data -  party nationalization dataset
  \item constituency.level.data - constituency nationalization dataset
}
}
\description{
This function builds the party nationalization datasets at three different levels
             of aggregation:  the national level, the party level, and the constituency level.
}
\examples{

library(CLEA)
library(readstata13)

# Open dat file to create ENP datasets
 dat<-read.dta13(
           unzip(
             system.file("clea_lc_20190617.zip", package="CLEA")
             ), convert.factors = FALSE)
 enp_data<-BuildENPdatasets(dat, filterIndependents=TRUE)


# Download file from http://www.electiondataarchive.org/ and create ENP datasets
 enp_data<-BuildENPdatasets(dat, filterSingleConstituencies=FALSE)

}
